Add Module metadata
This module is named specifically after the CVE identifier based on
early feedback from Randall and Luke that this should be as "one time"
use as possible.

The module may not require stdlib in the end.  I'd like for it to
require nothing but for now the facts we need to read certificate
information are hosted in stdlib.