Add "fix" script to secure a master.
This script takes a PE master and implements the steps required to
secure it.  A new certificate is generated using a new DNS name, a new
CA is generated, and the PE HTTPD configuration is configured to trust
both the new and the old CA.

After running this process, an existing Puppet Agent system should be
able to reconnect to the master by simply updating its --server setting.
e.g.

    puppet agent --test --server puppetmaster.new