Configure Apache to not peek into the CA directory
The Apache configuration should not use resource contained within the CA
directory because this directory is being rebuilt from scratch.

This change to the fix script ensure the Apache virtual host in PE is
configured to only use private keys and certificates from the
certificate directory itself, rather than the CA's "private" directory
where it does its thing.

This is really only "broken" for the CRL file.

In addition, the Apache virtual host is reconfigured to use the _new_
puppet master SSL certificate generated by the _old_ CA.  This ensure
existing agents are able to talk to the secured Puppet Master while they
are migrated to the new CA.

Finally, puppet agent --test is run on the Master to ensure MCollective
keys are regenerated for all of the agents.  Luckily, we have a
self-healing module as long as the master is configured with the
mcollectivepe class before the agents are.