Configure PE Apache to use explicit bundles
This change modifies the setup script to reconfigure PE Apache to use
explicit CA and CRL bundles rather than the same files used by the
Puppet Agent on the master node.

This is an effort to make it clear which authentication files Apache is
using as distinct from the authentication files the Puppet Agent is
using.