Add prototype migration module
This commit adds the prototype of the migration module.  If the
cve20113872 class is added to the catalog for a node, the following will
happen:

  * The catalog will configure localcacert and hostcrl to work with both
    the old and new certificate authorities.
  * The catalog will move the agent SSL directory out of the way.

As this is a prototype, we need to improve this module to work with FOSS
puppet.  It is currently hard-coded to work with Puppet Enterprise Agent
systems.  We need additional facts for the following configuration
settings:

  * localcacert
  * hostcrl
  * The CA currently in use on the Puppet Master