(WIP) Migrate the master SSL server cert to the new CA
This is a work in progress to perform the final step in the migration
process.  Once all of the agent systems have new SSL certificates issued
by the new CA, we need to migrate the Puppet Master.

 * Have a new certificate issued by the new CA.
 * Switch back to the original DNS name.

The Puppet Agent systems can only trust one CA at a time while
revocation checking is turned on.  They will be orphaned again and will
need to have two files deleted to reconnect to the new master:

  * localcacert
  * hostcrl