Replace use of Perl with Ruby for file patching
Nigel mentioned we may not be able to assume Perl for all of the
situations we're targeting.  We may assume Ruby, however, and at least
Ruby 1.8.7 for Puppet Enterprise customers.

This change replaces the perl -pli -e idiom with ruby -pli -e.  The
substitutions are accomplished using gsub with a block to evaluate the
back references.