Set ca_name in puppet.conf and allow existing agents to connect
The behavior without this patch applied is that once the master has been
migrated to listen on the new DNS name, existing agents could not
connect on the old name.

This patch generates a new master certificate using the new DNS name,
but adds the previous DNS name to the alternate names field.  This
enables existing agents to continue to use the old name.  This will
provide the opportunity to reconfigure the --server setting in puppet
conf to flip them to the new, secure, name.

In addition, the newly generated CA does not have its name added to
puppet.conf without this patch applied.  This patch will add the ca_name
setting to puppet.conf on the master to make it clear the master is
using a new, secure, CA name.