Make step3 names required arguments
The behavior before this patch is that the step3 script was incorrectly
assuming the DNS names being migrated from and to.

This patch makes the names required arguments.  This step performs the
work if migrating the Puppet Master to a SSL certificate issued by the
new CA containing the final name as the CN and the intermediate name as
the alternate name attribute.