Reload the agent when it generates a new CSR
Without this patch it takes two puppet runs on the agent to get a CSR
submitted.  The first run knocks out $ssldir and the second run
recreates it and submits a CSR.

This patch adds an agent_pid fact and configures the cve20113872 class
to send a HUP signal to the Puppet Agent after moving the ssldir out of
the way.  This causes the agent to reload itself, even if it is a
onetime agent.  The reload then causes the ssldir to be regenerated and
a new CSR to be submitted to the CA.