Make the agent work with both CA's (disable CRL checking)
An agent cannot work with more than one CA if the certificate_revocation
option is true, which is the default.  This patch makes sure to disable
certificate revocation during step2 when it obtains a new certificate
from the new CA.

The agent posses a localcacert file which contains the CA certificate of
both the old and new CA.  This allows it to communicate with both so
long as revocation checking is disabled.  However, the agent is still
"vulnerable" while both CA's are trusted IFF it is not using the new DNS
name.

In step3, we need to undo this behavior when the master migrates itself
to the new CA after all of the agents have been migrated.  We should
re-enable CRL checking, reconfigure the agent to talk to the old,
vulnerable DNS name, _AND_ no longer trust the old CA.  This will ensure
the agent is secure and will complete the migration process.