Make new dns name a mandatory argument for step1
Nigel pointed out in the code review that the new DNS name should be a
mandatory argument of step1.  This patch implements that suggestion.

In addition, the autosign disable section has been removed because
autosign is not actually a risk once the certdnsnames option has been
disabled on the master.