Tighten disable revocation script regexps
Nigel pointed out a number of loose regular expressions when modifying
puppet.conf on the agent to disable certificate revocation checking.

This patch tightens up the regular expressions.  It also takes into
account Nick's feedback that the [agent] section is actually [puppetd]
on older versions of Puppet.

Finally, the script made the poor assumption the puppet.conf was in
either /etc/puppet/puppet.conf or /etc/puppetlabs/puppet/puppet.conf
This may not be the case.  This patch also updates the behavior of the
migration assistant to use the $agent_config fact to obtain the full
path to the puppet.conf the agent is actually using.  This path is fed
as ARGV[0] to the revocation disable script through the configuration
catalog.