Allow the user to restore if they mess up step1
Nigel brought up the need to allow the user to run through step1
multiple times because it will be nearly impossible to make "perfect"
for all deployments in all situations.

This patch addresses this need by creating a backup copy of all the
files we modify in step1 before modifying them.  In addition, a pe
specific restore script allows the end user to restore to the state of
their system before doing anything related to this CVE migration
assistant.

The implementation assumes GNU tar because we only run PE master on
Linux systems.

This patch also addresses a loose regular expression in step1 and
tightens it up when modifying the ca_name setting.

Closes #2