Remove /vagrant paths and assume the module is installed first
The behavior without this patch is that step1 assumed it was running
from my personal development location of /vagrant/modules.

We're now going to assume the first thing a customer does is install
this module into their module path.  As a result, we no loner should try
and install the module in step2.  This part of step2 is removed by this
patch.

Furthermore, there is no longer a need to hard code any paths since it's
all either a single script or a puppet manifest itself.