Support node yamldir as well as facts
When the inventory service is enabled, we don't store facts in YAML,
which means we don't have the data necessary for this script to work.
Now, it can alternately be pointed at the node yamldir, and will produce
similar results.

Also added better error handling when there are no files, as well as
when a file can't be read. In the latter case, we now simply report
'unknown' for the issuer and timestamp, rather than failing the entire
run outright.