Make cve module self contained and not require stdlib
Without this patch, the cve20113872 module made use of the validate_re()
function from stdlib.  To minimize the number of steps the end user
needs to make this patch copies the function into this module.

This means stdlib is no longer required and the function has been
renamed to prevent conflicts with stdlib should it already be installed
which is the case on PE 1.2.3.

Closes #7