Add the cve20113872 class to top scope
Without this patch the cve20113872 class is added to node default which
may not result in all nodes having the class included in their
configuration catalog.

Based on feedback from Nick F, this patch changes the behavior to
instead add the class to the very top scope of site.pp outside of any
node definition.  This should result in every node having the class
added to its catalog.

Closes #9