Do not leave the master stopped
Without this patch, if the user runs through step1 twice, then we bail
out when we detect the backup copy.  We also leave their master shut
down, which isn't very nice.

This patch moves the chunk of code that stops the service after the
chunk of code that bails out if the backup file is present.

This results in us leaving their services the way they are if we refuse
to run a second time.