Make steps 1 and 2 only deal with securing the fleet
Without this patch, a new CA is generated in step1.  This new CA may
never get used if the user wishes to stop the process once their entire
fleet has been secured by switching to the new DNS name.

This patch addresses the problem by making some major changes to the way
the process is carried out.

First, steps 1 and 2 are only concerned with securing the Puppet nodes
themselves.  To this end, step1 simply issues a new SSL certificate for
the master with an additional subject alternative name that has not been
issued to anyone else.  Existing agents will continue to work after this
step, however they are still vulnerable until they use the new name.

Step 2 is strictly responsible for configuring all of the agent nodes to
use the new intermediate DNS name.  This is implemented as a Puppet
class named cve20113872::step2.  The step2 script simply adds this class
to the configuration catalog of all nodes by including it at top scope.

The generation of the new CA in step1 (without this patch) is moved into
a discrete step3 (with this patch).  Apache is also re-configured to use
the new CA bundle in step3, causing the master to authenticate agents
who posses certificates issued by both the old and new CA.

This patch _removes_ the functionality of causing agents to generate new
CSR's and submit them to the new CA.  This functionality will be added
back in as step4 as a subsequent patch.  The logic remains in the puppet
class cve20113872::step4 however.

This patch makes the README stale.  The README will be updated in a
subsequent patch.

Closes #18 - The intermediate certificate issued in step1 contains all
previous DNS names now.

Closes #19 - The cve20113872::step2 class is now included at top scope
not in the node scope.