Avoid modifying Apache configuration file
Without this patch, the Apache configuration file is being modified to
point at the bundle of CA certificates and CRL lists once a new CA is
generated.  The user will ultimately migrate solely to the new CA and it
may be confusing for them if we change the configuration half way
through the process then change it back at the end.

This patch avoids the need to modify Apache's configuration by appending
the old CA certificate to the newly generated CA certificate at the
location the Apache configuration points to by default in PE.