Add agent_server fact
While agents are being secured in step2, it will be nice if we provide a
way for the end user to check if each node is using the new, secure
intermediate DNS name, or a vulnerable name.

This patch adds the server configuration option to the list of facts the
agent reports to the master.  When the agent is flipped from one DNS
name to another, this may be monitored server side in
yamldir/node/*.yaml