Change server setting in puppet.conf even if it's absent
Without this patch, step2 makes the assumption that the server setting
is present in puppet.conf.  Nigel mentioned we need to take into account
the situation where the customer has removed the server setting and
relies on the default.

This patch takes this into account by using puppet agent --genconfig to
determine if the server setting is at its default value or not.  If it
is, then we look for a [main] section and add the server setting at the
top of the section.  If there is no [main] section then we simply add
one.  This should then work for puppetd and puppet agent.