Add agent cert reissue (CSR) back in as step 4
This patch adds the step to reissue a new SSL certificate to the agent
system.  This step was previous removed when refactoring out steps 1 and
2 to focus on securing the fleet only.

Step 4 will help customers clean up agent certificates which may
incorrectly have subject alternative names in them.  The clean up
process involves issuing new certificates from a brand new CA created
and configured in step 3.

This patch also introduced the behavior where we no longer turn off
certificate revocation checks on the agent.  We don't need to do this
anymore once this patch is applied because the catalog that triggers the
new CSR _also_ gives the agent the current CRL and CA certificate of the
new CA.  This means the agent will "just work" once the master is using
a certificate issued by the new CA.

However, this also means the node will refuse to talk to the master
until the master completes step 5.  In step 5 the master will replace
the SSL certificate issued by the previous CA with a SSL certificate
issued by the new CA.  The agents will no longer trust the previous CA
after step 4, causing them to be "disconnected."