Issue a new dashboard certificate in step3
Without this patch the dashboard inventory service and other things that
use the dashboard SSL client certificate no longer work after completing
step 5.  The reason is that the dashboard trusts only the previous CA
and the master trusts only the new CA.  The dashboard needs a client
certificate issued by the new CA.

This patch generates a new dashboard certificate if it detects the
dashboard directory on the same host as the Puppet Master.  It then
copies the private key, certificate, public key, and CA bundles into the
dashboard's certs/ directory.

Closes #33