(#36) Restore the agent's server setting in step 4
As nick points out in #36 the agent still uses the intermediate DNS name
configured in step 2 when it has completed step 4.  At the end of Step
4, the agent possesses a certificate issued by the new CA and only
trusts the new CA.  The server possesses a certificate that works on the
original DNS name.

This patch saves a backup copy of the puppet.conf file before modifying
it in step2 to contain the server = <intermediate dns name> setting.

The patch also updates step4 to restore this file if it exists, thereby
reverting the agent to its original configuration.

In order to avoid step 2 flipping to the intermediate name again after
step 4 flips it back this patch introduces a semaphore file.  The exec
resource that modifies puppet.conf will not run again if step 2 has been
completed with the application of this patch.

Closes #36