(#37) Restore site.pp to original state in step5
As Nick points out in #37, we should make a best effort to restore the
configuration catalogs of all the nodes to their original state before
starting the remedy process.

Without this patch, site.pp would still include the cve20113872::step2
and cve20113872::step4 classes in the top scope of the manifests.

This patch addresses the problem by saving an original copy of site.pp
in step2 before making any modifications.  This has the effect to
rolling the filebucket server setting back if it's set and removing the
cve20113872 classes.

Closes #37