(#24) Fix unrecognized escape sequences
Nigel points out in #24 the functions validating the values of the
custom facts this module uses throw warnings into the system log.  These
look like:

    warning: Unrecognised escape sequence '\.' in file
    /etc/puppetlabs/puppet/modules/cve20113872/manifests/init.pp at line 44
    warning: Unrecognised escape sequence '\.' in file
    /etc/puppetlabs/puppet/modules/cve20113872/manifests/init.pp at line 46
    warning: Unrecognised escape sequence '\.' in file
    /etc/puppetlabs/puppet/modules/cve20113872/manifests/init.pp at line 48

The problem is that the code uses double quotes which causes the Puppet
parser to perform string escaping before passing the string to the
underlying function.

This patch fixes the problem by converting the validation functions to
string literals, thus avoiding the string escapes and allowing the
underlying regular expression parser to handle the escape sequences
properly.

Closes #24