Fix tight puppet user group validation
Nigel pointed out in [1] that the regular expression validation for the
puppet user and the puppet group was too tight.  We support things other
than puppet and pe-puppet.

This patch relaxes the validation to ensure the string has a leading
word character and nothing else.  This will defend against an empty
string, which will be the most common problem and prevent false positive
failures with things like "root'

[1] https://github.com/puppetlabs/puppetlabs-cve20113872/blob/master/manifests/step2.pp#L30