(#42) Tighten the logic preventing step4 from applying to a master
Nigel points out in [1] the logic protecting the puppet master from
receiving the resources in step 4 is not as robust as it could be.  If
the PE facts are disabled for some reason or removed, then the resources
would still be applied to the puppet master.

This patch fixes the problem by adding a is_migration_host fact to the
module.  This fact will have three states we care about.  The empty
string means the fact didn't load or isn't present.  True means the
agent is running on the same host this migration is happening on.  False
means the fact correctly determined the agent is NOT the migration host.

Detection is done by looking for the dns_name file written in step1 to
$vardir on the master migration host.

In addition, Nigel verbally mentioned concern to me over the if
conditionals used in the Puppet manifests.  These may not be backwards
compatible if this module is used on very early versions of Puppet.
I've converted the if conditionals to a case statement, which has been
present in Puppet as far back as I can remember.

Closes #42

[1] https://github.com/puppetlabs/puppetlabs-cve20113872/blob/master/manifests/step4.pp#L80