(#31) Record the progress of the fleet on the master
Without this patch, we don't have an easy way to track the progress of
individual nodes as they work their way through the remediation process.
This patch adds a custom Puppet function to store when a node receives
the configuration for a particular state.

We're using a parser function because facts are submitted before the
configuration catalog is actually compiled.  Therefore, we'd need two
puppet run intervals to accurately determine the state.

The function approximates the state by recording when a configuration is
sent to the agent rather than reporting on the actual configuration
using facts.