(#31) Add check_progress script
This patch adds a check_progress script that provides information about
the progress of the fleet as it migrates through the remediation
process.  The output looks like:

    Status as of: 2011-10-23 11:27:02

        Total Nodes:              4 *
        Step0 (Has not run):      0 (  0%)
        Step2 (DNS Switch):       0 (  0%)
        Step4 (SSL Switch):       4 (100%)

     * Total of the nodes active within the last 30 days

Detailed information about the progress of each node is accessible using the
--yaml flag.  This will produce a hash table showing each node's status.

The progress is calculated by first looking at the $yamldir/node/ directory and
identifying all of the nodes who are active within the past 30 days.  A node is
considered to be active if the node cache file has a recent mtime.

In addition, the cve20113872_store_progress puppet function writes the progress
of each node as it receives the step2 and step4 puppet classes.  This progress
information is read from $yamldir/cve20113872/*.yaml

The script has no functionality to auto-refresh itself, but it is expected the
end user could run it in "watch" or a while loop on the shell.