(#52) Make remedy scripts support patched deployments
This patch adds support for Puppet versions that have the
--dns_alt_names option in addition to the --certdnsnames option.  This
is necessary if the user has already applied the CVE patch versions.

The change is fairly straight forward, simply using puppet master
--configprint dns_alt_names to determine if the dns_alt_names option is
valid.  Puppet correctly returns an exit code in this case.

The patch doesn't use grep to avoid SIGPIPE exceptions in the output
from Ruby when the pipe is closed.

Closes #52