(Maint) Add CVE_PATH check to run FOSS
Without this patch it is difficult to run these migration scripts with a
FOSS version of Puppet, or Puppet from source.  This patch reconfigures
all of the step script to use a CVE_PATH environment variable which
defaults to /opt/puppet/bin.  When running from source it enables
something like:

    CVE_PATH=/vagrant/bin ./pe_step1_enable_intermediate_dns_name

Which will use puppet in /vagrant/bin/puppet rather than
/opt/puppet/bin/puppet.