(#31) Improve progress check with certificates
Without this patch, it is possible for a node to be flagged as reaching
step 4 from the compiler function but it does not actually submit a CSR
because the agent is unable to obtain files.

This patch updates the check_progress script to add additional
information for nodes who progress beyond step 2.  These nodes are
placed in "buckets" depending on the status of their CSR request or
their signed certificate.

The patch also slightly changes the formatting of the previous script.
The new output looks like:

    Status as of: 2011-10-23 16:28:51

                                    Total Nodes:      4 *
                           Step 0 (Has not run):      0 (0.0%)
                            Step 2 (DNS Switch):      1 (25.0%)
                            Step 4 (SSL Switch):      3 (75.0%)

     * Total of the nodes active within the last 30 days

             Risk Mitigated (Issued a new Cert):      1 (25.0%)
                   Risk Mitigated (Pending CSR):      2 (50.0%)
            Risk Mitigated (Using new DNS name):      1 (25.0%)

Closes #31