(#49) Add scan_certs tool to identify SAN certificates
This patch adds a new utility names scan_certs to the cve module.  This
utility simply looks at all nodes in the $cadir/signed/ directory.  For
every *.pem file found here, we load the certificate and inspect if the
subjectAltName extension is present.  If so, the certificate is flagged
as potentially vulnerable.

By default, the script outputs a summary count of all the nodes who have
the SAN extension set in their certificate.  The output looks like:

    Status as of: 2011-10-23 19:42:26

                       Total Certificates Found:      7 *
                         Potentially Vulnerable:      7 (100.0%)

    * (Determined by looking at /etc/puppetlabs/puppet/ssl/ca/signed/*.pem)

    Potentially Vulnerable nodes are those who have the subjectAltName extension in
    their certificate.  The --yaml option to this script will provide detailed
    information

Closes #49