Make progress script work with FOSS
The progress script is hard coded to the ruby interpreter for PE because
we can't easily determine the bang line without a wrapper script.

This patch simply replaces #! /opt/puppet/bin/ruby with /usr/bin/env
ruby and then removes the checks that determine if the full path to
puppet is executable.

I've uncovered a ZeroDivisionError when this is run with no nodes in the
system.  This problem will be addressed in a subsequent patch to both
scripts.