(#52) Fix step 5 assumption of --certdnsnames
Without this patch, the patches added for #52 are not complete.  Step 5 still
calls `puppet cert --certdnsnames` without checking if PE has been upgraded to
the fixed version that ignores this option.

This patch updates the step 5 scripts to correct generate a puppet master
certificate using the --dns_alt_names option if the puppet master has already
been patched.

This fix does not add --dns_alt_names to the webrick (FOSS) scripts, but
it does add a defense mechanism in step1.  Issue #71 has been filed to
support --dns_alt_names in the FOSS remediation process.  The defense
mechanism introduced in this patch should be removed once issue #71 is
addressed.

Closes #52