(#71) Add support for --dns_alt_names to FOSS remedy scripts
Similar to #52 the FOSS remedy process scripts need to take into account
the situation where the customer has installed a patched version of
Puppet before starting the remediation process.

If they have, then the foss scripts do not behave as expected without
this patch because puppet will ignore the --certdnsnames option when
patched.

This patch fixes the problem by first checking if --dns_alt_names is a
valid option.  If it is, then it will attempt to use this option instead
of --certdnsnames.

It will also check for the situation where the customer has not changed
the certdnsnames configuration setting in puppet.conf to dns_alt_names
yet.  If so, it will read the value of certdnsnames and translate it to
an acceptable vale for dns_alt_names by replacing :'s with ,'s.

Closes #71